1. Created MyApplication class. This will be used to store global variables (like DataGrabber) so we don't have to parcel things around so much. I have not implemented this yet. 2. Inexplicably, the force close that used to occur when changing screen orientation on ClassSwipeActivity does not happen any more. However, the class order gets shuffled. I'm not sure what's causing this. 3. Class averages are now taken from the description page and not calculated by hand (some of the averages were wrong, inexplicably). Average displays on each tab.

+++ src/com/sunstreaks/mypisd/net/DataGrabber.java
@@ -605,7 +605,7 @@ public class DataGrabber implements Parcelable {
 		Object[] termCategory = Parser.termCategoryGrades(html);
 		JSONArray termCategoryGrades = (JSONArray) termCategory[0];
 		
-		if ((Double)termCategory[1] != -1)
+		if ((Integer)termCategory[1] != -1)
 			classGrade.getJSONArray("terms").getJSONObject(termIndex).put("average", termCategory[1]);
 		classGrade.getJSONArray("terms").getJSONObject(termIndex).put("grades", termGrades);
 		classGrade.getJSONArray("terms").getJSONObject(termIndex).put("categoryGrades", termCategoryGrades);
